refactor: use bloc builder to show user info and renamed profile_2 to sign_up
