+++ .dart_tool/version
@@ -1,1 +1,1 @@
-2.2.3
+2.5.0-5.2.pre
