+++ .flutter-plugins
@@ -1,4 +1,4 @@
 # This is a generated file; do not edit or check into version control.
-flutter_plugin_android_lifecycle=/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.3/
-image_picker=/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/
-image_picker_for_web=/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.3/
+flutter_plugin_android_lifecycle=/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.3/
+image_picker=/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/
+image_picker_for_web=/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.3/
