+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"image_picker","path":"/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.3/","dependencies":[]},{"name":"image_picker","path":"/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/","dependencies":["flutter_plugin_android_lifecycle"]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"image_picker_for_web","path":"/Users/oliwiabrewko/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.3/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"image_picker","dependencies":["flutter_plugin_android_lifecycle","image_picker_for_web"]},{"name":"image_picker_for_web","dependencies":[]}],"date_created":"2021-10-18 11:31:47.821968","version":"2.2.3"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"image_picker","path":"/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.3/","dependencies":[]},{"name":"image_picker","path":"/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/image_picker-0.8.4+1/","dependencies":["flutter_plugin_android_lifecycle"]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"image_picker_for_web","path":"/Users/marcossevilla/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.3/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"image_picker","dependencies":["flutter_plugin_android_lifecycle","image_picker_for_web"]},{"name":"image_picker_for_web","dependencies":[]}],"date_created":"2021-10-18 10:00:03.875519","version":"2.5.0-5.2.pre"}
