+++ lib/app_navigator.dart
@@ -1,4 +1,4 @@
-import 'package:app/profile_1/profile_view.dart';
+import 'package:app/sign_up/sign_up_view.dart';
 import 'package:flutter/material.dart';
 
 class AppNavigator extends StatelessWidget {
@@ -8,7 +8,7 @@ class AppNavigator extends StatelessWidget {
     return Navigator(
       pages: [
         // Show loading screen
-        if (isCurrentUserSelected = true) MaterialPage(child: ProfileView())
+        if (isCurrentUserSelected = true) MaterialPage(child: SignUpView())
 
         // Show auth flow
       ],
